Refactor toast and auth context, fix lint
