+++ project-bolt-sb1-qao21pjs/project/packages/design-system/src/index.ts
@@ -2,7 +2,8 @@
 export { Button } from './components/Button';
 export { Card } from './components/Card';
 export { Modal } from './components/Modal';
-export { Toast, ToastProvider, useToast } from './components/Toast';
+export { ToastProvider } from './components/Toast';
+export { useToast } from './components/ToastContext';
 export { Input } from './components/Input';
 export { Spinner } from './components/Spinner';
 export { Avatar } from './components/Avatar';
@@ -16,4 +17,4 @@ export type { ButtonProps } from './components/Button';
 export type { CardProps } from './components/Card';
 export type { ModalProps } from './components/Modal';
 export type { InputProps } from './components/Input';
-export type { ToastType } from './components/Toast';
+export type { ToastType } from './components/ToastContext';
