+++ project-bolt-sb1-qao21pjs/project/packages/portal/src/features/auth/AuthContext.tsx
@@ -1,0 +1,28 @@
+import { createContext, useContext } from 'react';
+
+export interface User {
+  id: string;
+  email: string;
+  role: 'student' | 'faculty' | 'admin';
+  name?: string;
+  avatar?: string;
+}
+
+export interface AuthContextType {
+  user: User | null;
+  loading: boolean;
+  login: (email: string, password: string) => Promise<void>;
+  logout: () => void;
+  refreshToken: () => Promise<void>;
+}
+
+export const AuthContext = createContext<AuthContextType | null>(null);
+
+export const useAuth = () => {
+  const context = useContext(AuthContext);
+  if (!context) {
+    throw new Error('useAuth must be used within an AuthProvider');
+  }
+  return context;
+};
+
